backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx==0.26.0

# AI/ML
openai==1.8.0

# Utilities
python-dateutil==2.8.2

# Environment
python-dotenv==1.0.0